(PUP-10368) Validate file against desired checksum
The `content_checksum` is the checksum of the temporary file we just wrote.
Previously, we were computing the same checksum and comparing it against
`content_checksum` which always matches. This was originally added in
b06767ee2d7c22c27d746d3e8d1b6effa37deaa6 because puppet would sometimes corrupt
file content.

The `checksum_value` property was added in 99c254dbf011ed3a9085771ad18a10faed24418a
so that the desired checksum value could be specified with the `source`
parameter. But puppet didn't verify that the file it wrote matched that
checksum, which is a bit surprising.

If the checksum type is neither none or time based and the checksum type and
value are specified, then verify the temporary file's checksum matches
before the real file is replaced.