(PUP-10572) Create loaders in the agent application
Puppet failed to load cached catalogs containing rich data, because the
Configurer hadn't yet created the loaders necessary to deserialize the catalog
using pcore. This affected rich data types like Binary and Deferred, but not
Sensitive, since it predates pcore and is serialized differently.

This commit moves the creation of the loaders to the `agent` application similar
to how the `apply` and `device` applications already handle it. See commits
b9878fd0258 and b30ecfcc429.

If the environment changes based on the node response, then update the current
environment like we did before. Note there is a separate ticket to update the
current environment if the environment changes due to catalog convergence, see
PUP-10308.

This commit also preserves creation of the loaders if it hasn't been done
before requesting a catalog. That part should be removed in puppet 7.