(PUP-10368) Test checksum validation when using content or source
Verify checksum validation when file content is specified using the `content`
property or `source` parameter. In the later case, the insync-ness is handled
by the `checksum_value` property.

The `content` cases incorrectly set the content to `file contents`, but the hash
values were generated for an empty file.

The checksum mismatch cases are pending due to PUP-10368.