(PUP-10580) Preserve URL query when using absolute form
The legacy Connection class allows the caller to specify a URL whose
scheme, host and port are different than the Connection was created with. The
Connection then passes the entire URL in 'absolute-form' on the request line.
The ConnectionAdapter class did not preserve query parameters in that situation.

Update the `to_url` method to preserve the query parameters. The authinfo and
fragment components are intentionally not preserved.