(PUP-10580) Don't double encode path and query
Previously, the Connection and ConnectionAdapter encoded the URL path and query
differently. The Connection class required callers to encode them, while the
ConnectionAdapter assumed the caller hadn't. As a result, any code calling the
legacy HTTP API could end up with double encoded paths and queries.

The ConnectionAdapter was made to encode the URL path and query because webmock
made it seem like the Connection class did. The following stub:

    stub_request(..).with(query: "foo=bar%3Dbaz")

matches the unescaped query:

    http.get("..?foo=bar=baz")

So it appeared the Connection class performed one level of encoding. However,
the stub matches the following query too:

    http.get("..?foo=bar%3Dbaz")

So it seems webmock matches too leniently and masks the Connection class' actual
behavior.

This commit removes the URL encoding from the ConnectionAdapter so it matches
the Connection class, adds more realistic JSON query parameters to the tests and
adds an integration test.