(PUP-9602) Pass pcore resource type loader
Previously the Runtime3TypeLoader created a pcore resource type loader solely
based on the `env_path` being nil or not. However, this means we're creating a
pcore resource type loader for `puppet agent` and during the catalog application
phase of `puppet apply`. In order to make this conditional, create the pcore
resource type loader in the Loaders factory method, and pass it to the
Runtime3TypeLoader.