(PUP-9602) Verify loader hierarchy
Add tests for the loader hierarchy in the default case, when tasks are enabled,
and `for_agent`. In the default case, the hierarchy is:

    static loader
    puppet system loader
    runtime 3x type loader
      delegates to pcore resource type loader
    public environment loader
      delegates to private environment loader

where each loader refers back to its parent, except for static which is the root
of the hierarchy. Note the 'puppet system' and public environment loader are
both instances of LibRootedFileBased, so verify both the loader name and class.

When tasks are enabled, the runtime 3x type loader is excluded, as well as its
pcore resource type loader delegate.

When `for_agent` is enabled, the `cached_puppet_lib` loader is added after the
`puppet_system` loader, but before the runtime 3x type loader.