(PUP-9602) Refactor loader creation
Add private method for creating loaders for cases where puppet apply is given a
manifest or an already compiled catalog.

The apply application appeared to create one set of loaders to compile the
catalog and another to apply it. However, the `apply_environment` object never
changes, and the `Puppet::Pops::Loaders.new` method will (surprisingly!!) return
the existing set of loaders from the environment instead of creating a new one.
So remove the second `Puppet.override` which doesn't do anything.