(PUP-10547) Manage user rights on Windows
This feature allows puppet to manage the rights and privileges of
Windows users through the existing `roles` property. The accepted values
are, as described in Microsoft documentation, certain constant names
(https://docs.microsoft.com/en-us/windows/security/threat-protection/
security-policy-settings/user-rights-assignment) which can be passed in
an array or as a string when it is just one.

This property comes in pair with the `role_membership` parameter which
offers better control over the desired state through the two supported
values: `minimum` (default) and `inclusive`. When this parameter is set
to `minimum`, the given roles are appended to the existing ones. When
it is set to `inclusive`, managed user remains only with the given
roles. The `roles` property set to an empty array in combination with
the `role_membership` set to `inclusive` can be used to remove all roles
of an user.

AD Users are currently still not supported for manangement with Puppet.