(PUP-2608) Return default file metadatas for builtin mounts
Previously the agent printed an ugly warning if it tried to pluginsync, but the
server did not have any top-level module directories. The puppetserver packages
create several default module directories like `/etc/puppetlabs/code/modules`
so this doesn't often occur, though it used to occur with a newly installed
webrick puppet master in 4.x, and can occur if the directories are deleted.

The motivation for this change is because the server's file metadata response
should differentiate between the set of file metadata is empty versus missing.
The agent should fail in the latter case, but not the former.

This commit changes the `search` method for plugins,pluginfacts,locales mount
points to return the default path Puppet[:codedir] so that the file server
returns FileMetadata for that directory, which the agent will apply to its
:plugindest,:pluginfactsdest,localedest puppet settings, respectively.

Also remove warning, which is logged at debug and not very useful.