(PUP-10599) skip local external facts on lookup
Some puppet default settings required facter to retrieve the data which
caused facter to load all external facts that are not used. This can
slow down puppet if external facts are slow.

Facter.load_external(<boolean>) is set to false when using Facter to
retrieve a core fact that does not require loading external facts.
After the value is retrieved the flag is set to True so it works as
before.