(maint) Clear registered termini instances after each test
Preserve the `@termini` hash before running each test so that it is restored at
the end. We have to duplicate the hash so that the saved copy isn't modified
during the test. The hash maps registered termini names (eg :processor) to
objects for each indirection (eg :report).

Prior to this, tests could reuse a terminus created and registered during an
earlier test. If the terminus' initialize method had side effects, like creating
a settings catalog, then those side effects could be skipped, causing the test
to fail, but not when run in isolation.

We can't use the terminus_class reader because it will raise if the variable
is currently nil and there isn't a setting to specify the default terminus class.