(maint) Return code_content that matches checksum
The static catalog beaker test failed, because as of commit 039270e298, puppet
will raise if the file it writes doesn't match the desired checksum. For static
catalogs, the desired checksum is specified in the `metadata` section of the
catalog. The test didn't fail previously, because puppet would write the file,
and then read the file back in, and compare those two checksums, ignoring the
desired checksum in catalog.