(PUP-10585) Allow transaction store to YAML load Time
Commit 0b6563eea09 allowed Puppet::Util::Storage to load Symbol and Time from
yaml, but only added Symbol to Puppet::Transaction::Persistence. As a result,
puppet would fail to load the transactionstore if it had stored a Time in the
prior run, issue a error and prevent puppet from reporting on corrective change.

Add Time to the allowed list of classes and add a test like the "should load
Time and Symbols" test in storage_spec.