(PUP-10587) Add plan_hierarchy key to Hiera config
This adds a new `plan_hierarchy` key to Hiera's config at the same level
as the `hierarchy` key. This key is used to specify data to use for
lookups in Bolt plans outside of [apply
blocks](https://puppet.com/docs/bolt/latest/applying_manifest_blocks.html).

This solves two related problems with lookups in Bolt plans:
1. Bolt plans don't run in a context that has a node, environment, or
   other common interpolations used when configuring and looking up
   Hiera data, which causes
2. Users who reuse their existing Hiera configuration with Bolt often
   encounter errors because interpolation isn't supported, even if the
   data they are trying to look up doesn't require interpolating.

Adding the `plan_hierarchy` key allows users to add to their existing
Hiera config and use existing Hiera data without error, and clarifies to
users what data can be used in Bolt plans (outside apply blocks) vs.
what data can be used in Puppet (and apply blocks).

The plan_hierarchy configuration will be used when Pal is using it's
`ScriptCompiler`, the compiler used to compile plans. This doesn't
include apply blocks, which are compiled by the `CatalogCompiler`. The
new key should not have any impact on existing Hiera configuration
files, with the exception that files that include the `plan_hierarchy`
key will not raise a validation error even if they are not used in Bolt
context. This is a good thing though, since we expect users to use the
same Hiera config for both Bolt and non-Bolt operations. Other than
that, existing lookups should have the exact same behavior.

The new key is only supported in Hiera 5.