(maint) Add `Logon As A Service Right` logic for PUP-1289
This commit completes the implementation of the `Manage service logon
credentials on Windows` feature (PUP-1289) by adding logic based on
given user rights and better tests of the expected behaviour.

When using a Windows user as logon account for a service, it is
important to consider the rights/privileges it has. Focused on in this
implementation is the `Logon As A Service` right, identified and used
with the constants `SeServiceLogonRight` and `SeDenyServiceLogonRight`.

When managing `logonaccount` property, given user will now be checked if
it has the `Logon As A Service` right and fail when missing. The same
happens when it is denied. Puppet will raise an error regarding this
before any changes to the `logonaccount` parameter occurs. Services that
have such user set as logon account cannot start as it will result in a
logon failure. These rights can be managed accordingly using Puppet but
only for local (non-AD) users for now.

Above scenarios and when user password changes between puppet runs got
covered in the acceptance tests.