(PUP-1763) Stop puppet run if pluginsync fails
If an error occurs during pluginsync and `Puppet[:ignore_plugin_errors]` is
false, then stop the run, otherwise facts needed to compile the catalog and
types/providers needed to apply the catalog will be missing. The lack of facts
can cause compilation to fail, and the lack of type & providers can cause
application to fail, or worse data corruption due to mismatched versions of old
plugins and new catalogs or vice versa. To ensure backwards compatibility, the
setting defaults to true, so there is no change in current behavior, though the
default value will be changed to false in Puppet 7.

Part of the reason for the old behavior is that 4.x and earlier webrick-based
puppet master replied with HTTP 404 if it didn't have any modules installed with
at least one `lib` directory OR if the requested environment didn't exist. The
former is not an error, while the latter is. Had we changed the agent's behavior
earlier, then agents would fail until at least one module was installed on the
server.

Puppetserver 5.0 and above replies with HTTP 404 if none of the top-level
`modules` directories exist such as `/etc/puppetlabs/code/modules`, but it no
longer requires at least one module with a `lib` directory. The puppetserver
package creates several `modules` directories during installation, so HTTP 404
should only occur if the requested environment doesn't exist on the server.

This commit updates the downloader so it doesn't rescue errors that occur during
pluginsync for facts, lib and locales. This does mean that the agent will fail
fast during environment convergence, so it will never attempt to request a
catalog from a non-existent environment. This isn't an issue, because previously
the server rejected the agent's request in the indirected routes handler, before
the ENC had a chance to influence the agent's environment.