(maint) Use block form of Puppet.debug
Prefer the block form of `Puppet.debug` when passing an interpolated message, so
that the interpolation can be avoided when running in non-debug. This is similar
to commit 2d0d7333b, but includes server-side puppet code that is called on
every request, such as features, confines, gettext, network formats and route
matching. It also updates file system and Windows code paths that are triggered
for every file resource.