(PUP-9602) Ignore pcore resource loader for puppet apply
Previously, `puppet apply` failed if the manifest referenced a resource default,
and there existed pcore resource types for that resource in
<env>/.resource_types/<type>.pp. The resulting error was:

    Could not autoload puppet/type/applytest: Attempt to redefine entity

During compilation the Runtime3TypeLoader delegated to the pcore loader to load
the resource type from the <type>.pp file, and cached the resulting
PResourceType object in the Runtime3TypeLoader's cache.

Later during catalog compilation, the `Catalog#to_ral` method called the
metatype manager to autoload the resource type and register it with the
Runtime3TypeLoader. However, due to the already registered PResourceType object,
the above error was raised.

To fix this issue we could include the pcore loader during compilation, but
exclude it from catalog application. However, creating two hierarchies is slow.
Instead this commit modifies the apply application to disable the pcore loader,
so that the '<type>.pp' files are ignored. The autoloader will load types during
catalog compilation and those types will be reused during catalog application,
which is the same way puppet used to behave before the pcore loader was
introduced.