(PUP-10252) Remove VerifierAdapter
The `HttpPool.http_ssl_instance` method accepted an instance of
`Puppet::SSL::Validator` and passed it as the `verify` argument to the
`Connection#initialize` method. The `http_ssl_instance` method was deleted in
PUP-10439 and the connection factory methods don't accept a Validator as an
argument. So it's no longer possible for the Connection class to receive a
Validator when using the legacy HTTP API.

This commit deletes the VerifierAdapter and modifies the Connection class so
that it no longer accepts a `verify` option.