(PUP-10252) Remove Puppet::SSL::Host
The Host logic for bootstrapping the key/csr/cert was buggy and extremely
complicated due to the way the DefaultValidator#setup_connection method could
cause network connections while in the middle of configuring a network
connection. The Host logic for key/csr/cert parsing and persistence is
superseded by Puppet::X509::CertProvider, and the logic for building an
OpenSSL::X509::Store is superseded by Puppet::SSL::SSLProvider.

This commit deletes the Puppet::SSL::Host class and removes the `:ssl_host`
object from the context system.