(PUP-10252) Remove ssl_{server,client}_ca_auth
The `ssl_client_ca_auth` setting was an allow-list of root or intermediate CA
certificates that the agent would trust when connecting to a server. The setting
has been broken for along time and is no longer used.

The `ssl_server_ca_auth` setting was used in the ruby puppet webrick/passenger
server, but has been dead for a long time.

This commit removes both settings and the ExternalCertFixtures helper that
referenced the settings but is no longer used in puppet beaker tests.
Puppetserver provides its own helper when testing external certs.