(PUP-10440) Remove networking code from indirector request
The indirector request object contained logic about how to serialize and encode
HTTP parameters and SRV resolution, even though that is specific to HTTP/REST
termini, but not memory, file, etc termini.

All of the HTTP serialization/encoding logic is in the HTTP client, so remove
it from the indirector request.