(PUP-10603) Consistently encode metadata and content for http URLs
Prior to 6.16, puppet assumed http file sources in manifests were already URL
encoded. So to reference a URL path with a space, you had to specify:

    source => 'http://example.com/b%20c'

Similarly, if the query contained a character that wasn't a URL code point[1],
such as 'b[c', then it had to be URL encoded in the manifest:

    source => 'http://example.com?q=b%5Bc'

Commit 11eee4e886d changed the http content request so it called the `uri`
reader method instead of calling `URI(metadata.source)`. The former encodes the
source using `Puppet::Util.uri_encode` before parsing it, which could result in
double encoded source URLs. It also meant the http content used different levels
of encoding than the previous http metadata request.

The `uri_encode` method also incorrectly assumes `=` characters are query
delimiters, which may not be the case for a base64 encoded signature with
trailing padding characters like

    sig=JaZhcqxT4akJcOwUdUGrQB2m1geUoh89iL8WMag8a8c=

This commit restores the previous behavior of using `URI(metadata.source)`
instead of the `uri` reader. Doing so bypasses the `=` issue in query strings,
which is filed as PUP-10613 and will be fixed later, and ensures the metadata
and content requests use the same levels of encoding.

[1] https://url.spec.whatwg.org/#url-code-points