(PUP-10615) Add facts and vars back to ScriptCompiler scope
This reverts the changes made in puppetlabs/puppet#8063, which
removed facts and vars from PAL's ScriptCompiler scope. This
resulted in a breaking change that affected downstream projects
that rely on PAL's public API.

It also adds a new `:set_local_facts` option to `#with_script_compiler`,
which will toggle facts being set for the compiler scope. The
default value for the option is `true` to preserve the old behavior of
the script compiler, while allowing other tools (namely Bolt) the
ability to control the facts variables being set.