(PUP-10618) Use load_as_module? to load Bolt project
Previously, Bolt would only set the `bolt_project` Puppet setting if the
project was loadable, in this case if it had a name. However Bolt always
has a project, it just may not get loaded as a module. This presented
issues when we needed to know where the Bolt project was located in
Bolt's custom Puppet functions, but the Bolt project wasn't named. The
interim solution for this was to set a new Puppet setting
`bolt_project_data` to read from the custom Puppet functions. This
introduces a more robust solution, which is to only load the project if
a method `load_as_module?` returns true. This lets us always set the
`bolt_project` Puppet setting for access throughout the Bolt lifecycle,
while only loading the project as a module when valid.

See https://github.com/puppetlabs/bolt/issues/2051 for more details.