(PUP-9405) Mark resources as failed for LoadError and MissingCommand
Puppet defers evaluation of a resource until its provider becomes suitable or
the provider is the default. If the provider implements prefetch, then it will
be called to determine the current state of all resources of that type.

Previously, if prefetch raised LoadError or MissingCommand, then the resources
associated with the provider were not marked as failed. Instead, puppet assumed
no resources of that type existed, and puppet would call the provider to create
them(!), possibly corrupting the system such as occurred in PUP-2643 with
crontab entries.

This commit modifies prefetch so if the method raises LoadError or
MissingCommand, then all associated resources are marked as failed.