(PUP-10656) Provide a JSON fact terminus
We have a YAML terminus for facts, however YAML can problematic for hexidecimal numbers (see PUP-9505).
Additionally, JSON is generally a better serialization format for machines to understand, is a smaller spec than YAML, has a relatively mature ecosystem in Java around parsing/serializing, and is backwards compatible with YAML for recent YAML parsers (YAML has become a superset of JSON).
For these reasons we should implement a JSON terminus for facts (there already exists one for catalogs) and users should be encouraged to use it instead of the YAML terminus.