(PUP-10617) Call puppetserver simple status endpoint
Resolve server_list by trying to connect to the puppetserver's simple status
endpoint for each entry in the list. If the connection succeeds, then return a
service for that name, host and port. Since resolution uses the puppetserver
service, then puppet related HTTP headers are included in the request, unlike
requests to arbitrary HTTP servers.