(PUP-10628) Short circuit TypeSetType inference
The TypeSetType defines a namespace for a set of PCore types. Internally, puppet
uses the TypeSetType when registering all of the AST model classes. See
Puppet::Pops::Mode.register_pcore_types. Puppet also uses the TypeSetType class
when autogenerating ast.rb from the PCore model defined in ast.pp. See the
generate_ast_model rake task.

The Puppet::Pops::Lookup::DataProvider verifies external sources return data are
confined to RichData types, such as string, boolean, etc. In commit 0a195ef6d6,
the RichData type alias was defined as:

    Variant[Scalar,SemVerRange,Binary,Sensitive,Type,TypeSet,Undef,Hash[RichDataKey,RichData],Array[RichData]]

The 'TypeSet' above refers to the PTypeSetType class. The commit also changed
the DataProvider to accept all RichData[1].

Since RichData is an alias for the Variant above, the DataProvider calls
Variant#instance? passing the data to be validated. The Variant type delegates
to Scalar, SemVerRange, etc checking to see if the data is an instance of that
type.

The PTypeSetType inherits its instance? implementation from its parent
PMetaType. That version naively uses the TypeCalculator to infer the value of
the data passed to it. When the data is a Hash or Array, the TypeCalculor must
infer the type of every element recursively, which can be extremely expensive
for large collections.

This commit implements PTypeSetType#instance? and short-circuits the check if
the object is not a Hash or is missing the 'name' or 'pcore_version' keys. In
practice the other keys like 'name_authority' are omitted, so we can't require
them. Doing so results in 300x speed up as described in PUP-10628.

[1] https://github.com/puppetlabs/puppet/commit/0a195ef6d6c1791ed45b64abb59f1374619a7049#diff-e2d2e02e4e793dc7d69cbb05b1042024L13-R17