(PUP-10628) Check for an instance of PTypeSetType
Check that the object passed to `PTypeSetType#instance?` is an instance of
`PTypeSetType` or subclass, instead of a `Hash` with a pcore_uri key. This
redoes the change from commit 786e8b366.

The `PTypeType#instance?` method already checks for specific classes, so doesn't
cause full type inference of a Hash.

Add the `PTypeType` and `PTypeSetType` to the list of rich data compatible
classes, so they are tested in the type calculator specs. Those classes
correspond to the `Type` and `TypeSet` in the `RichData` alias.