(maint) Handle slow 2019 appveyor workers
For tests where the foreground agent tries to acquire the lock and exits with
failure, increase the timeout that the background thread waits for the
foreground agent to start, as it might take awhile in appveyor.

For the test where the background thread acquires the lock and waits for the
foreground agent to fail to acquire the lock once, have the foreground agent
signal the failure using the log destination. As soon as the failure occurs, the
background thread will exit, enabling the foreground agent to acquire the lock
and successfully apply the catalog.