(PUP-10473) Ignore authorization rules in custom fileserver mounts
Previously, puppet parsed `allow` and `deny` rules in legacy auth.conf,
but did not support trying to restrict access to something other than
"allow all". Doing so would log an error and deny access. So for
example, `allow *.example.com` would always be denied.

This commit removes the error case. It is assumed authorization is performed
outside of puppet.

This commit preserves denying access if the requested fileserver mount
does not exist.