(PUP-10664) Log server_list exceptions as they occur
Previously, the server list resolver collected exceptions that occurred during
resolution. Now we log the exceptions at error level as they occur, which
eliminates the need for the `error_handler`.

The server list resolver also used to raise an error if the list was exhausted,
but that prevented the collected exceptions above from being logged at error
level. Now the server list resolver uses the `canceled_handler` to cancel
resolution (instead of raising) and its resolve method returns nil like other
resolvers.

The integration test relied on enabling debug for server_list tests. When that
is removed, it causes puppet to print the error to stderr instead of stdout.