(PUP-10677) gem provider uses system ruby
Because `INSTALL_DIR/puppet/bin` is prepended to PATH,
gem provider always uses Puppet’s Ruby.
This commit changes the gem provider to execute commands
in a specific environment that has a PATH which does not
contain `puppet/bin`.

It also updates the puppet_gem provider to use
the full path for the Puppet’s gem executable