(PUP-8014) Touch environments as they are retrieved from the cache
Add a touch method to the base Entry class, and call the method whenever the
entry is loaded from the cache. Also log a debug message when the cache is used.
The message is passed as a block so that the string is not allocated or
interpolated when not in debug mode.

The touch method is empty because the existing set of entries either expire
always, never or at an absolute time (ttl seconds after the entry was created).