(PUP-8014) Cache most recently used environments
If `environment_ttl` is explicitly set, then cache the environment for up to
that many seconds since it was last created or retrieved from the cache. If the
setting is not explicitly set, then use `environment_timeout` as before.

Note even if `environment_ttl` is set and an environment is still active, the
environment may still be expired via the `cache_expiration_service`.

Using the global `environment_ttl` setting means we don't need to read
`environment.conf` from the environment's directory (to determine what it's
environment_timeout is). This won't solve the issues described in PUP-6408, but
it will reduce file IO whenever an environment is loaded and cached.