(PUP-8014) Return environment_ttl in environments REST API
If the `environment_ttl` setting is explicitly configured, and therefore
preferred over `environment_timeout`, then return that value in the
`/puppet/v3/environments` REST API.

The `environment_ttl` and `environment_timeout` settings behave the same for
0 (never cache) and 'unlimited' (always cache). They have different semantics
when using values in between, as the former is the ttl since it was last
accessed, while the former is a ttl since it was created.

However, `environment_timeout` with values other than 0 or 'unlimited' is
deprecated, and the setting will be removed in Puppet 7. So send the
`environment_ttl` value in the `environment_timeout` field of the response.