(PUP-8014) Call cache expiration service with symbolic environment name
Previously, puppet called the cache expiration service's `expired?` method with
the environment name as a Symbol, but called the `evicted` method with a String.
It appears to work because of some JRuby interop magic.

Use Symbols when calling `evicted` and update the cache service yardocs.

Also split the `custom cache expiration service` test into the difference cases
where an environment is created, is not expired and is expired. Also change the
ReplyExpirationService so that the individual test can set expectations directly
on the service rather than buried in the `expiration_sequence`.