(PUP-10670) Create `serverport` setting and alias
This commit creates a `serverport` setting which serves the same
functionality as `masterport`. Update the puppet code itself to use the
new setting, and add a hook to the old one to always set the new one as
well.

Due to how puppet interpolates settings, we cannot use integer values as
defaults, so just set both masterport and serverport to the same value
to ensure nothing breaks.