(DO NOT MERGE) POC for async scopes in Bolt
Bolt is looking to add asynchronous blocks to plans, which means that
using the existing scopes won't work. The current Parser::Scope creates
a 'global' table of variables, and tracks how many variables existed
when a new block begins. When the block finishes it 'pops' the list back
down to the original number when the block was created. This works
great for linear execution, but when blocks execute concurrently it
means that the number of variables created during the block may not be
equal to the difference between the global variables at the beginning
and end of the block. That's because variables from other blocks will
also get added and removed from the global scope while the block is
executing, modifying the length of the global scope. This also leaks
variables since variables defined in 'inner' blocks will be available
globally to outer blocks, since everything uses the same variable list
`@ephemeral`.

This PR is a proof-of-concept for a new evaluation model which will
create a new Parser::Scope object for each block and add any existing
variables from the outer scope to the new scope. This ensures that each
block has a unique scope that doesn't know about any other scopes, so
tracking variables is unnecessary. This isn't a fully formed solution,
as cleaning up scopes once the blocks have finished is unhandled and
there may be other issues with this approach.