(PUP-9505) Quote facts that contain special string
Description of the problem: puppet facts --render-as yaml outputs mac addresses unquoted.
If the mac address doesn't contain any chars A-F this is a problem as other tools parsing the resulted yaml will interpret the value as a sexagesimal (base 60) integer.

Description of the fix: use Psych to double quote any string that contains the : character (this affects macaddress, any date/time and ipv6 style values).