(PUP-10669) Server settings section
As master section is replaced with server section, for now
master section is still allowed in puppet.conf, but when
adding/updating a setting from the master section, the action
will be done on the server section instead, and the setting from [master]
will be removed.

Also when deleting a setting from either [master] section or [server] section
it will delete the setting from both [server] and [master] sections.