(PUP-10598) Update expectation when puppet agent fails
Prior to PUP-10598, puppet would ignore the pluginsync errors due to the
non-existent environment on the server and would fail when trying to request a
catalog.

As of PUP-10598, puppet will fail on the first pluginsync error, so the error
message is different. The overall behavior is similar though. The cached catalog
is ignored since `puppet agent --test` implies `--no-usecacheonfailure`.