(PUP-10671) Update setting section search path
Puppet adds a section to its searchpath based on the run mode. This is
what enables puppetserver running in `:master` run mode to load settings
from the `[master]` section. As a result of the switch to `:server` run
mode, then applications running in server run mode, like `puppet
lookup`, will no longer read settings from `[master]`.

If our run mode is `:server` or `:master`, add sections for both, in
that order, so that `:server` takes precedence over `:master`.

Add `server` as an allowed section name to be able to test this change,
and update the remaining `master` run mode tests to use `server`.