(PUP-10473) Verify puppetserver registers its authconfigloader
Verify puppetserver registered its authconfigloader class, which indicates it
will authorize REST requests. If it fails to do so, either due to
misconfiguration or bug, then return HTTP Unauthorized (401) for all v3 REST
endpoints so we fail securely.

Note when legacy routes was enabled (in 6.x) indirected routes performed the
authorization check within its `uri2indirection2` method. But here we do the
check before the route is called, since we just need to know the request method
and path to format the error message.