(PUP-8014) Add environment_timeout_mode
The `environment_ttl` setting was added to replace `environment_timeout`, but
having two settings complicates the upgrade process for the common case where
the timeout is set to 'unlimited'. If we naively upgrade, then puppet server
will switch from always cache to never cache, due to the default value of
`environment_ttl=0`.

This commit preserves the existing behavior of `environment_timeout` when the
value is `0` or `unlimited`.

Otherwise, the `environment_timeout` value is the time to live relative to when
the environment was created or last used. The `environment_timeout_mode` setting
defaults to `from_created`, but can be set to `from_last_used`.