(PUP-2173) Validate setting before committing changes
Previously, it was possible to configure the `certname` value to an invalid
value, which prevented the agent from starting the next time it reloaded its
config.

This commit sets the value programmatically and then looks up the value. It is
not sufficient to just set the value, since we could be setting an
uninterpolated value that references a non-existent setting, such as `puppet
config set path '$condir', and it's only when we lookup the value that
interpolation and `munge` validation is performed.

If the setting isn't a value that puppet recognizes, such as if you want to set
a setting prior to upgrading puppet, then puppet will behave as before.

If the setting is a value that puppet recognizes, but the enumerated value is
unknown, such as digest_algorithm=rot13, then validation will fail.