(maint) Remove ignored paths from GitHub Actions
Due to having all GitHub Actions jobs marked as required in the
repository's GitHub Branch settings, we always had pull requests with
stuck jobs waiting to be run when there were no code changes (no changes
at all in cases like merge ups or only in `**.md` or `docs/*` or
`exampes/*` or `ext/*` or `benchmarks/*` files). Workaround for this was
to have someone with enough rights on the repository to forcefully merge
the pull request.

This commit makes all the jobs run always even if the files added/modified
in the pull request can't impact the code in any way.